Backend/Backend/requirements.txt
django>=3.2.14
djangorestframework
django-cors-headers
django-allauth
dj-rest-auth
djangorestframework-simplejwt
python-dotenv
gunicorn
